+++ src/application/position.hpp
@@ -26,31 +26,31 @@ namespace noteahead {
 struct Position
 {
     Q_GADGET
-    Q_PROPERTY(uint32_t pattern MEMBER pattern)
-    Q_PROPERTY(uint32_t track MEMBER track)
-    Q_PROPERTY(uint32_t column MEMBER column)
-    Q_PROPERTY(uint32_t line MEMBER line)
-    Q_PROPERTY(uint32_t lineColumn MEMBER lineColumn)
+    Q_PROPERTY(size_t pattern MEMBER pattern)
+    Q_PROPERTY(size_t track MEMBER track)
+    Q_PROPERTY(size_t column MEMBER column)
+    Q_PROPERTY(size_t line MEMBER line)
+    Q_PROPERTY(size_t lineColumn MEMBER lineColumn)
 
 public:
     //! Index of the pattern.
-    uint32_t pattern = 0;
+    size_t pattern = 0;
 
     //! Index of the active track of the pattern.
-    uint32_t track = 0;
+    size_t track = 0;
 
     //! Index of the active note column of the track.
-    uint32_t column = 0;
+    size_t column = 0;
 
     //! Index of the active line (common for all tracks).
-    uint32_t line = 0;
+    size_t line = 0;
 
     //! Index of the active column of the line consisting of note and velocity values.
     //! 0: Note, e.g. "C-3"
     //! 1: Velocity digit 2, e.g. [1]27
     //! 2: Velocity digit 1, e.g. 1[2]7
     //! 3: Velocity digit 0, e.g. 12[7]
-    uint32_t lineColumn = 0;
+    size_t lineColumn = 0;
 
     bool operator==(const Position & other) const
     {
